[moleculer] refac based on @icebob 's feedback
