+++ src/extension/logs/wts.log..2019-10-09
@@ -1,0 +1,2 @@
+[2019-10-09T14:01:38.223] [INFO] default - [ EXTENSION ]  Launched
+[2019-10-09T14:01:39.481] [INFO] default - [ EXTENSION ]  Successfully synced templates. Version: 0.0.0.0
