--- templates/Web/Projects/MoleculerDefault/services/pages.actions.js
@@ -1,4 +1,0 @@
-const SERVICE_ACTIONS = {};
-SERVICE_ACTIONS.actions = {};
-
-module.exports = SERVICE_ACTIONS;
