+++ templates/Web/_composition/MoleculerJS/Page.Moleculer.AddSampleData.ForAllExceptList/services/pages.service_postaction.js
@@ -1,5 +1,6 @@
+"use strict";
 //{[{
+
 const sampleData = require("../data/sampleData");
 //}]}
-
-const SERVICE_ACTIONS = {};
+module.exports = {
