+++ templates/Web/_composition/MoleculerJS/Page.Moleculer.Grid.AddRoutes/services/pages.service_postaction.js
@@ -1,0 +1,15 @@
+  /**
+   * Service Actions
+   * More info: https://moleculer.services/docs/0.13/actions.html
+   */
+  actions: {
+    // Action handlers
+  //{[{
+    grid: {
+      rest: "GET /grid",
+      handler(ctx) {
+        return sampleData.textAssets;
+      }
+    },
+  //}]}
+  },
