--- templates/Web/_composition/MoleculerJS/Page.Moleculer.MasterDetail.AddRoutes/services/pages.actions_postaction.js
@@ -1,13 +1,0 @@
-const SERVICE_ACTIONS = {};
-SERVICE_ACTIONS.actions = {};
-
-//{[{
-SERVICE_ACTIONS.actions.masterDetail = {
-  rest: "GET /masterdetail",
-  handler(ctx) {
-    return sampleData.textAssets;
-  }
-};
-//}]}
-
-module.exports = SERVICE_ACTIONS;
