+++ templates/Web/_composition/MoleculerJS/Page.Moleculer.MasterDetail.AddRoutes/services/pages.service_postaction.js
@@ -1,0 +1,15 @@
+  /**
+   * Service Actions
+   * More info: https://moleculer.services/docs/0.13/actions.html
+   */
+  actions: {
+    // Action handlers
+  //{[{
+    masterDetail: {
+      rest: "GET /masterdetail",
+      handler(ctx) {
+        return sampleData.textAssets;
+      }
+    },
+  //}]}
+  },
