+++ templates/Web/_composition/MoleculerJS/Page.Moleculer.StartConfig.NoCosmos/services/pages.service_postaction.js
@@ -10,9 +10,8 @@
   /**
    * Service Mixin
    * More info: https://moleculer.services/docs/0.13/services.html#Mixins
-   * This will extend (current) `pages` service with action handlers.
    */
-  mixins: [SERVICE_ACTIONS],
+  mixins: [],
 //}]}
   /**
    * Service dependencies
